Switched from using google/dense_hash to std::unordered_map for better performance
Originally Google dense hash type was brought in since it provided a
reasonably good speedup when fairly large contouring tasks were
profiled. However, now that we're moving to truly huge datasets in the
order of tens of gigabytes, the number of edges in contours has
increased so much that it has become obvious Google dense hash has its
limits, and most likely was not designed for such a large number
of elements.